board vo Authority 삭제 Dao user 통합 로그인 타입 영문으로 변경 home 유저 권한설정 오류메시지 추가
