+++ src/main/java/dao/UserDao.java
@@ -2,6 +2,8 @@ package dao;
 
 import util.DaoHelper;
 import vo.Gym;
+import vo.Membership;
+import vo.Trainer;
 import vo.User;
 
 public class UserDao {
@@ -72,5 +74,26 @@ public class UserDao {
 			user.setGym(gym);
 			return user;
 		}, email);
+	}
+	
+	public void insertMembership(Membership membership){
+		DaoHelper.update("membershipDao.insertMembership", membership.getCat(), 
+														   membership.getType(),
+														   membership.getName(),
+														   membership.getDescription(),
+														   membership.getPrice(),
+														   membership.getDiscountedRate(), 
+														   membership.getNumOfUseDay(),
+														   membership.getNumOfUseWeek(),
+														   membership.getNumOfPause(),
+														   membership.getDuration(),
+														   membership.getCount(),
+														   membership.getGym().getNo());
+		
+	}
+	public void insertTrainer(Trainer trainer) {
+		DaoHelper.update("trainerDao.insertTrainer", trainer.getUser().getId(),
+															trainer.getPosition());
+			
 	}
 }
