+++ src/main/webapp/manager/form.jsp
@@ -45,7 +45,7 @@
 }
 %>				
    			<form class="border bg-light p-3" method="post" action="insert.jsp">
-   				<input type="hidden" name="type" value="관리자">
+   				<input type="hidden" name="type" value="manager">
    				<div class="form-group mb-2 w-75">
    					<label class="form-label">아이디</label>
    					<input type="text" class="form-control" name="id" />
